add sleep for script acme
Former-commit-id: 0aaadc2da1c90f932445581c0d9798b4faca82fe

+++ Packer/start.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 echo "Start"
+sleep 30s
 ### waiting to be exist file
-while [ ! -f "/dev/disk/by-id/scsi-0Google_PersistentDisk_persistent-disk-1" ]; # true if /your/file does not exist
-do
-  sleep 1
- echo "wait file exist"
-done
-echo "file already exists, continuing"
+if [ ! -f "/dev/disk/by-id/scsi-0Google_PersistentDisk_persistent-disk-1" ]; # true if /your/file does not exist
+then
+echo "file exists, continuing"
 sudo mkdir /mnt/traefik-acme
 sudo mkfs.ext4 -F -E lazy_itable_init=0,lazy_journal_init=0,discard /dev/disk/by-id/scsi-0Google_PersistentDisk_persistent-disk-1
 sudo mount -o discard,defaults /dev/disk/by-id/scsi-0Google_PersistentDisk_persistent-disk-1 /mnt/traefik-acme
 sudo touch /mnt/traefik-acme/acme.json
 sudo chmod 600 /mnt/traefik-acme/acme.json
 sudo umount /mnt/traefik-acme
+else
+echo "file  not exists, continuing"
+fi
